squash! a27d3491575e8671facb76905fc132a95834d3aa
MDEV-38521 TO_DATE: Date strings with non-ASCII symbols cause warnings
           and wrong results

Known limitations:
- Format string cannot contain non ascii characters. This is because
  TO_DATE depends on the format string parser 'parse_format_string()'
  used by TO_STR(). Should be fixed at some point in time.

+++ strings/ctype-bin.c
@@ -92,7 +92,7 @@ static int my_strnncoll_binary(CHARSET_INFO * cs __attribute__((unused)),
   {
     if (cmp)
       return cmp;
-    return slen - tlen;
+    return (int) (slen - tlen);
   }
   if (cmp)
   {
@@ -165,7 +165,7 @@ static int my_strnncoll_8bit_bin(CHARSET_INFO * cs __attribute__((unused)),
   {
     if (cmp)
       return cmp;
-    return slen - tlen;
+    return (int) (slen - tlen);
   }
   if (cmp)
   {
